requirements.txt
flask==1.0
boto3==1.9.86
flask_admin==1.5.3
Flask-SQLAlchemy==2.3.1
Flask-Bootstrap4==4.0.2
Flask-Assets==0.12
pyScss==1.3.5
